Moved SLS services into services folder. Moved shared functions into common folder.
