+++ sls/common/response_utils.py
@@ -1,0 +1,21 @@
+import json
+
+def make_200(response_body):
+    return {
+        'statusCode': 200,
+        'headers': {
+            'Content-Type': 'application/json',
+            'Access-Control-Allow-Origin': '*'
+        },
+        'body': json.dumps(response_body)
+    }
+
+def make_500(error_message):
+    return {
+        'statusCode': 500,
+        'headers': {
+            'Content-Type': 'application/json',
+            'Access-Control-Allow-Origin': '*'
+        },
+        'body': json.dumps({ 'error': error_message })
+    }
