+++ sls/services/web-api/src/.gitignore
